Feat: Add macro for GET authenticated endpoint operations.
