+++ src/endpoints/mod.rs
@@ -63,7 +63,7 @@
 //!     // Get character research agents for character_id using the access_token
 //!     let research_agents = esi_client
 //!         .character()
-//!         .get_agents_research(character_id, access_token)
+//!         .get_agents_research(&access_token, character_id)
 //!         .await
 //!         .unwrap();
 //! }
